api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0
pydantic-settings==2.6.1
pydantic[email]==2.10.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
anthropic==0.40.0
pymupdf==1.24.0
python-docx==1.1.2
weasyprint==62.3
python-multipart==0.0.20
stripe==11.1.0
razorpay==1.4.2
celery[redis]==5.4.0
sentry-sdk[fastapi]==2.19.0
